requirements.txt
aws-cdk.core>=1.133.0
aws-cdk.assertions>=1.133.0
aws-cdk.aws-ecs>=1.133.0
aws-cdk.aws-ec2>=1.133.0
aws-cdk.aws-rds>=1.133.0
aws-cdk.aws-secretsmanager>=1.133.0
aws-cdk.aws-elasticache>=1.133.0
aws-cdk.aws-elasticloadbalancingv2>=1.133.0